Adjust frontend for basic names testing.
Minimal changes to the front-end template in order to make it useful
for testing the names pallet:  Add a second input field for the
"extrinsics" UI section, so that we can call the name update/transfer
extrinsics from it.  Also add the types needed to properly display events
and storage reads.