Implement configurable fee for names.
Make names cost a fee for registration and update.  The concrete fee for
an operation must be configured through the pallet's Trait (and may e.g.
depend on the name length).  Also what happens with the fee needs to be
configured on the Trait (e.g. burn it, distribute it to miners, send to
a fixed "root" account, ...).