Configurable name policy.
Change get_name_fee in the names trait, so that it can also return None
in case an operation is deemed invalid because e.g. the name is too short
or the value invalid.

With this, developers using the names pallet get the option to define their
own policy not just for fees, but for validation as well.